+++ tests/servers/storagetcp.py
@@ -17,20 +17,37 @@ class GetTokenMessage(Message):
     pass
 
 @dataclass
+class PutGlobalMessage(Message):
+    secret: str = ""
+
+@dataclass
+class GetGlobalMessage(Message):
+    pass
+
+@dataclass
 class TokenResult(Data):
     token: str = ""
+
+@dataclass
+class SecretResult(Data):
+    secret: str = ""
 
 
 class StorageProtocol(PlankyProtocol):
     async def parse_message(self, data: bytes):
         if data[0] == 0x00: return PutTokenMessage(token=data[1:].decode(errors="ignore"))
         if data[0] == 0x01: return GetTokenMessage()
+        if data[0] == 0x02: return PutGlobalMessage(secret=data[1:].decode(errors="ignore"))
+        if data[0] == 0x03: return GetGlobalMessage()
 
         return await super().parse_message(data)
 
     async def pack_message(self, message: Data) -> bytes:
         if isinstance(message, TokenResult):
             return bytes([0x01]) + message.token.encode(errors="ignore")
+
+        if isinstance(message, SecretResult):
+            return bytes([0x03]) + message.secret.encode(errors="ignore")
 
         return b""
 
@@ -48,6 +65,19 @@ async def get_token(client: PlankyClient, event: MessageEvent):
 
     token = client.storage.get("token", default="NoToken")
     await client.send_data(TokenResult(token=token))
+
+@server.on_message(PutGlobalMessage)
+async def put_global(client: PlankyClient, event: MessageEvent):
+    event.message: PutGlobalMessage
+
+    server.storage.set("secret", event.message.secret)
+
+@server.on_message(GetGlobalMessage)
+async def get_global(client: PlankyClient, event: MessageEvent):
+    event.message: GetGlobalMessage
+
+    secret = server.storage.get("secret", default="NoSecret")
+    await client.send_data(SecretResult(secret=secret))
 
 def mainloop():
     server.mainloop()