+++ docs/changelog.md
@@ -10,6 +10,7 @@
 - Added a "ping_time" to client to store the timestamp of the last ping.
 - "parse_extra" now don't return Extra class
 - add default value for Storage.get
+- Global storage for server
 
 ### Version 0.0.1.2